refactor: update home page open event functionality to follow listener/handler pattern
